we need to load the dna in shredded mode.  trying to insert a blob larger than ~200MB into a single field via DBI's $sth->execute() method causes an "Out of memory!" error.
the workaround here is to append 100MB chunks at a time by issuing an
UPDATE call on the target field.

on top of this, the postgres append function, || (good symbol to overload,
eh?), silently fails when it tries to append onto a NULL value.  so we
must first initialize the field to insert sequence into as an empty string.

this was a heinous bug to fix.